requirements.txt
streamlit
audio_recorder_streamlit
groq
elevenlabs
python-dotenv